Update patch set 1
Patch Set 1:

This might not be the right fix. In RTCEAGLVideoView, there is a RTCEAGLVideoViewDelegate which notifies when the video frame changes - 
- (void)videoView:(RTCEAGLVideoView *)videoView didChangeVideoSize:(CGSize)size;

For RTCMTLVideoView, there isn't. So we can't know when to crop/fill the video view based on the incoming video.

We should fix RTCMTLVideoView so that we get notified of the size change and can resize/crop the view. Or RTCMTLVideoView needs to handle the video size/device screen differences.

Patch-set: 1
Reviewer: Gerrit User 5590 <5590@58829da1-049c-39fb-b951-ebdcd0984767>