Merge to M66: Changing MTU size for SCTP socket options.
With the latest usrsctp roll, the MTU value you provide is the space
avaiable for chunks in the packet. We previously specified this to be the
MTU for the entire SCTP packet, so we were logging errors when the SCTP
packets were 12 bytes larger than expected (the size of the SCTP header).
This fix updates our MTU specified to account for the SCTP header size
as well.

TBR=shampson@webrtc.org

Bug: webrtc:9082
Change-Id: I9ceb1d0cc5bee5fbad76844f732d9f09ad48da1d
(cherry picked from commit 5150ee40f45fe9cdef03b4454bfde7e099860673)