Update patch set 4
Patch Set 4:

(2 comments)

Nice work!

What are the expected values of S and N in the VAD? If S is much higher that N there will be a lot of copying and a circular buffer would be preferred.

Please, also have a look at the in-line comments.

Patch-set: 4
Reviewer: Gerrit User 5116 <5116@58829da1-049c-39fb-b951-ebdcd0984767>