Update patch set 3
Patch Set 3:

If I understand correctly, the approach is to break out the RuntimeSetting class from the audio_processing.h header file for the purpose of allowing RuntimeSetting objects to be directly passed into the AecDump class.

While I see that it simplifies the code for writing, I think it complicates the use of the AudioProcessing interface and at least will require some significant name changes to make sense.

The reason is that this CL moves the RuntimeSetting from being a nested class in the AudioProcessing class to being a standalone class which means that the context of the RuntimeSetting class is changed from belonging to AudioProcessing to being standalone. 
This is kind of ok now, but once audio_processing.h is moved to api/audio (or similar) runtime_setting.h will also need to be moved there. Then, since the nested context within AudioProcessing is lost, name/namespace changes will be needed to specify that the class is not a generic RuntimeSetting but a RuntimeSetting tailored only to AudioProcessing.

I don't think the code simplifications for recording AECdumps achieved by this approach are worth making the APM API surface/usage more complicated by separating RuntimeSetting from AudioProcessing.

Patch-set: 3
Reviewer: Gerrit User 5125 <5125@58829da1-049c-39fb-b951-ebdcd0984767>