Injection of network simulation into FakeNetworkPipe.
This CL adds the functionality to inject custom FakeNetworkInterface
implementations into FakeNetworkPipe. A side effect of this is that
the SetConfig function on FakeNetworkPipe no longer can be used,
since custom implementations can have their own configuration format.

In the places where SetConfig was used, a SimulatedNetwork instance
is now injected, on which SetConfig can be called.

Bug: None
Change-Id: I3a0d15535419408355fce9d5c2551fca409772ec