Update patch set 1
Patch Set 1:

Is there any update on this CL? I see crash in the same place inside our app and this CL will probably fix it. It would be nice if it will be accepted. Maybe I can somehow help, for example providing necessary changes to make this CL acceptable if it has been abandoned by author?

Patch-set: 1
CC: Gerrit User 6305 <6305@58829da1-049c-39fb-b951-ebdcd0984767>