Detect network interface types in Linux using ioctl.
A new method of detecting network interface types is added and used as
the first round of detection in BasicNetworkManager, before the existing
approach using name matching in GetAdapterTypeFromName. This method is
added to increase the reliability of WebRTC in detecting the correct
adapter type and consequently assigning correct network costs to
candidates.

Bug: None
Change-Id: I22f6ee337c5d1292187e745d752c61b303d40213