[desktopCapture] Unify the position info in DIP coordinates on Mac.
On OSX, the logical(DIP) and physical coordinates are used mixingly.
For example, the captured image has its size in physical pixels(2x) and
location in logical(DIP) pixels. Same to the cursor position. This
causes trouble when we check the relative position of image and cursor
when there are multiple monitors with different DIP setting connected.

This cl proposed a solution to use DIP pixel for any location info,
i.e. top-left of a frame and cursor position, together with current
scale factor on Mac and add a method to provide rect size in DIP.
Then we can check their relative position correctly in DIP pixel and
compose them in physical pixel.

Notes: There are some obselated codes. Will keep them until the
new methods work totally fine.

Bug: webrtc:9178
Change-Id: I3c076aeac2d6f2c1f63d000d7fff03500aa375ac