Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Actually instruction is a bit dangerous, because in such case it will check in latest revision for new dep and won't update other deps, which can lead to some misses. Better is to get 'chromium_third_party_revision' from DEPS file and run roll_deps.py with --third-party-revision <chromium_third_party_revision>. It will checkout new dep on the same revision that have others deps. Also it good to pass --revision <chromium_revision>.

I don't see why that is dangerous? We're bringing a new dep into existence, so there can't be anything in WebRTC that depends on a particular version of it. Or, well, it can happen that the new dep in turn depends on a certain version of something in third_party.

I actually wrote the instructions to pass --revision and --third-party-revision first, but it seemed kind of useless to me since the script would still roll other things (like chromium/base), and I would still have to git checkout -- DEPS.

I think what we should do is drop the --revision and --third-party-revision flags and introduce a new flag --in-place which doesn't update the hashes but just does the copying.

Patch-set: 3
Reviewer: Gerrit User 5027 <5027@58829da1-049c-39fb-b951-ebdcd0984767>