Update patch set 5
Patch Set 5:

(5 comments)

> Patch Set 4:
> 
> (1 comment)
> 
> Could you please elaborate a bit more on why merging the interfaces is needed and how much redundant code is needed?
> 
> As I remember it, at the point that was written there was before a clear intention to have the aec_dump interface separate from the audio_processing interface and as you point out there are some clear advantages with that.

I and alessiob@ got to the conclusions that the benefits are small. The current interface structure is from an early iteration of AecDump. At first, we had a no-op AecDump impl which was instanciated from AudioProcessingImpl. Then having a target for the Aec Dump interface was more motivated.

Patch-set: 5