Update patch set 7
Patch Set 7:

(5 comments)

Hey Taylor, thanks for the thorough review of the interface! I agree using the signal-slot model gives much better decoupling between the two objects, and this may actually also address the other question you have on how we solve cases like BasicPortAllocatorSession::OnNetworkFailed in our current code. Right now we have this method directly reacts the signal from the network manager (we wired them), and I think if we want to reinforce the model that self-contained tasks in the ICE transport are done by different modules, and the ICE transport could act as the hub of signal/information, we could also move wire signals from the network manager to ICE, and let ICE relay the signals. This is not done yet in this CL but would be done in this one or a different CL if you think it makes sense. This structure may become a pattern we use for the following refactoring and new modules to make the ICE transport more modular. Thoughts?

I feel introducing the controller factory may need another iteration since it may be correlated to how we design the ICE public API, and keep it on my wish list. Also, I have merged your comments for the controller, and they are really good!

Patch-set: 7
Reviewer: Gerrit User 5887 <5887@58829da1-049c-39fb-b951-ebdcd0984767>