Update patch set 7
Patch Set 7:

(11 comments)

> Patch Set 7:
> 
> (5 comments)
> 
> Right now we have this method directly reacts the signal from the network manager (we wired them), and I think if we want to reinforce the model that self-contained tasks in the ICE transport are done by different modules, and the ICE transport could act as the hub of signal/information, we could also move wire signals from the network manager to ICE, and let ICE relay the signals. This is not done yet in this CL but would be done in this one or a different CL if you think it makes sense. This structure may become a pattern we use for the following refactoring and new modules to make the ICE transport more modular. Thoughts?

Hmm. Making one "ICE transport" interface the hub for everything may insulate us from refactoring. For example, depending on PortAllocatorSession in a bunch of places may not be a good idea if it's going to be replaced by something different; when it's replaced, everything that depends on it will need to be updated. But if everything depends on P2PTransportChannel instead, then only P2PTransportChannel would need to be updated.

But I don't think that's the best long-term design. It would lead to a bloated interface with poor encapsulation, and probably a lot of boilerplate code doing nothing but forwarding signals/methods to other classes. As long as there's a clear hierarchy of dependencies (no cycles), and we believe the modules will stick around for a while, I don't see a problem with modules depending directly on each other.

For example, this would seem like a good design to me:

            |IceAgent|
                |
                V
|----------IceTransport----------|
  |           |               |
  |           V               |
  |  |RegatheringController|  |
  |      |    |       |       |
  |      |    |       V       V
  |      |    | |IceStatsInterface|
  V      V    |
|IceGatherer| |
       |      |
       V      V
   |NetworkManager|

Where:

* IceAgent is the "brains" that selects a candidate pair.
* IceTransport handles connectivity checks, etc.
* IceStatsInterface is a pure virtual interface that provides stats to be used by other objects, implemented by IceTransport.
* IceGatherer just gathers candidates.
* RegatheringController uses IceStatsInterface to get stats, then uses IceGatherer to gather additional candidates.

Patch-set: 7
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>