Update patch set 3
Patch Set 3:

(3 comments)

I think makes sense.
One thing though, SpatialLayer has a member bool active that specifies if the stream is active and sent or not.
I don't think that info is used for vp9 svc, ssilkin@ can you comment?
If it is, make sure to use it. If it is not, perhaps add a comment that we need to account for that flag when calculating the padding bitrate.

Patch-set: 3
Reviewer: Gerrit User 5524 <5524@58829da1-049c-39fb-b951-ebdcd0984767>