Update patch set 13
Patch Set 13:

PTAL

> Patch Set 9:
> 
> This is a good step! I think we can polish it a bit:
> 
> 1. api/call/ should actually be removed (it's a relic from when we intended to expose
>    Call as a lower-level API for clients). Let's not add more to it.

Ok! I added bitrate_constraints.h to api/ instead. I merged media_type.h into mediatypes.h. This highlights the silly situation that we have both cricket::MediaType and webrtc::MediaType, but it's at least made more explicit now. I didn't think it made sense to have a mediatypes.h and a media_type.h in the same dir for sure.

> 2. I'm not sure I see the point of exposing VideoQualityTestFixture as an interface.
>    The tests using it follow the pattern:
>      create fixture
>      set up params
>      single call to RunWithAnalyzer or RunWithRenderers
>      exit
>    There are no iterative calls to a RunTest() method like for the VideoCodecTestFixture
>    test cases, and thus no need to maintain state -> we only need to expose the Run...
>    functions. Perhaps there is more functionality planned?

The downstream test needs to get the payload_type_map() out from the fixture currently, and it will probably need to talk a lot more to the fixture, so I think we're going to need it. The fixture is very stateful already, but hopefully we can only expose a minimal part of it.

> 3. simulated_network.h should live under api/test/, and the class SimulatedNetwork looks
>    ripe for hiding behind a Create function (the interface is already there!), but maybe
>    you intended that for a follow-up?

Moved it. I only needed simulated_network for the SimulatedNetwork::Config, which is used in the test params. Actually, maybe it's enough to just yank out that config struct and put back simulated network where it was.

Patch-set: 13
Reviewer: Gerrit User 5027 <5027@58829da1-049c-39fb-b951-ebdcd0984767>