Update patch set 9
Patch Set 9:

(10 comments)

>             |IceAgent|
>                 |
>                 V
> |----------IceTransport----------|
>   |           |               |
>   |           V               |
>   |  |RegatheringController|  |
>   |      |    |       |       |
>   |      |    |       V       V
>   |      |    | |IceStatsInterface|
>   V      V    |
> |IceGatherer| |
>        |      |
>        V      V
>    |NetworkManager|
> 
> Where:
> 
> * IceAgent is the "brains" that selects a candidate pair.
> * IceTransport handles connectivity checks, etc.
> * IceStatsInterface is a pure virtual interface that provides stats to be used by other objects, implemented by IceTransport.
> * IceGatherer just gathers candidates.
> * RegatheringController uses IceStatsInterface to get stats, then uses IceGatherer to gather additional candidates.

This design looks good to me, and it looks like the new IceTransport would be the low-level ICE that Peter proposed. We will get very close to it after we also refactor the port allocator to the gatherer (and replace the external dependency on it by a direct injection of the packet socket factory by the way, as one item on our hotlist). I am wondering if the IceStatsInterface would be part of layer that connects with the IceAgent instead of being under the hood of IceTransport. It feels ideal that the agent is the plug-and-play part by only interacting with the stats/states/signals from the IceTransport.

Patch-set: 9
Reviewer: Gerrit User 5887 <5887@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0