Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> I was looking at "expectations of owners" in https://chromium.googlesource.com/chromium/src/+/lkcr/docs/code_reviews.md, where some of the requirements are:

Oops, hit enter on accident. As I was saying, some of the requirements are:

- Be already acting as an owner, providing high-quality reviews and design feedback
- Be a Chromium project member with full commit access of at least 6 months tenure.
- Have submitted a substantial number of non-trivial changes to the affected directory.

If we're using these guidelines for WebRTC as well, then it may be a little premature to add you as an owner. It may be a good idea to get some experience doing code reviews for a while first.

Which I think you could do by adding yourself to https://cs.chromium.org/chromium/src/third_party/webrtc/WATCHLISTS, so that you're CC'ed on changes and can provide unsolicited reviews.

Question for other owners: Do we follow the chromium guidelines or are we less strict? What should the path for becoming an owner look like?

Patch-set: 3