Update patch set 5
Patch Set 5:

In the CL description where it today says: "Bug: None", please change it to "Bug: webrtc:xxxx", where xxx is a reference to a number in https://bugs.chromium.org/p/webrtc/issues/list where the background of why this change is needed is discussed. As is, I can't see which client would be the end user of this change, and if that is unknown, we can't guarantee that the code is not broken going forward.

Patch-set: 5
Reviewer: Gerrit User 5091 <5091@58829da1-049c-39fb-b951-ebdcd0984767>