Update patch set 3
Patch Set 3:

(2 comments)

My comments are mostly on the threading model of TurnServer and TestTurnServer. We may do something very similar to that in PortAllocator in thread checking: 1) allowing the object created (by DetachFromThread) and configured on a different thread; 2) having a milestone marked by some method like Init() and reinforce the single-thread model after an invocation of this method. This may help us avoid the thread hopping for ctors and setters.

Patch-set: 3
Reviewer: Gerrit User 6646 <6646@58829da1-049c-39fb-b951-ebdcd0984767>