Update patch set 11
Patch Set 11:

I'm not sure it's a good idea to add a new init function that only requires a custom audio device module and sets all other components to default values. The problem is that we are trying to reach a state where the core WebRTC API is as slim as possible and has no dependencies to e.g. SW audio encoders or similar. We want clients to be responsible of injecting all components they need, and if they don't inject e.g. audio and video they will get a datachannel only PeerConnection without any other dependencies. Adding helper functions like the one you have added in this CL goes against this and will make it harder for us to reach this state. Another problem is that the number of helper functions we need grows combinatorially with the number of injectable components, and it would be better with a Builder pattern instead like we have for Android (https://cs.chromium.org/chromium/src/third_party/webrtc/sdk/android/api/org/webrtc/PeerConnectionFactory.java?rcl=b90e63c620877712e45ee320cfa25cb825bf5373&l=134).

I'm fine with the other changes in this CL, but if you want to add a new init function in RTCPeerConnectionFactory+Native.h, do it with a Builder pattern. The default values in this Builder should be null, i.e. no dependencies to anything, but you can add a helper function in a separate file in a separate target that can initialize the Builder to contain default components. For your use case, you would call this default Builder, pulling in all dependencies, and then override the AudioDeviceModule to what you want.

Patch-set: 11
Reviewer: Gerrit User 5098 <5098@58829da1-049c-39fb-b951-ebdcd0984767>