Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 3:
> 
> (2 comments)
> 
> My comments are mostly on the threading model of TurnServer and TestTurnServer. We may do something very similar to that in PortAllocator in thread checking: 1) allowing the object created (by DetachFromThread) and configured on a different thread; 2) having a milestone marked by some method like Init() and reinforce the single-thread model after an invocation of this method. This may help us avoid the thread hopping for ctors and setters.

I was considering this as well, although this model felt a bit of an overkill for a simple test object like this. A compromise could be calling DetachFromThread in the ctor, allowing construction on a separate thread, but enforcing the threading elsewhere. One thing worth noting is that the ctor in the TestTurnServer creates a TurnServer, adds a socket for the server to listen on and then sets some configurations to the server. I don't know the details of sigslot/C++ in this case, but if the TurnServer immediately starts reading packets coming from the networking thread before the TestTurnServer ctor finishes it doesn't seem very thread safe. That's why I enforced the ctor called on the network thread as well. 

Let me know if you think it's worth adding this threading model you suggested, or just ping me if you want to talk about it. I kept it as is for now.

Patch-set: 3