Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> (2 comments)
> 
> > Patch Set 3:
> > 
> > (2 comments)
> > 
> > My comments are mostly on the threading model of TurnServer and TestTurnServer. We may do something very similar to that in PortAllocator in thread checking: 1) allowing the object created (by DetachFromThread) and configured on a different thread; 2) having a milestone marked by some method like Init() and reinforce the single-thread model after an invocation of this method. This may help us avoid the thread hopping for ctors and setters.
> 
> I was considering this as well, although this model felt a bit of an overkill for a simple test object like this. A compromise could be calling DetachFromThread in the ctor, allowing construction on a separate thread, but enforcing the threading elsewhere. One thing worth noting is that the ctor in the TestTurnServer creates a TurnServer, adds a socket for the server to listen on and then sets some configurations to the server. I don't know the details of sigslot/C++ in this case, but if the TurnServer immediately starts reading packets coming from the networking thread before the TestTurnServer ctor finishes it doesn't seem very thread safe. That's why I enforced the ctor called on the network thread as well. 
> 
> Let me know if you think it's worth adding this threading model you suggested, or just ping me if you want to talk about it. I kept it as is for now.

Just throwing in my opinion here.

I think in general we should avoid initialize methods if possible and have our classes do all initialization in the constructor (or a factory method/class if initialization could fail). Having an Initialize method means there's an awkward uninitialized state which is easy to forget about and not well tested.

With that, I think we should also avoid the pattern of "create on one thread then initialize on another". To me as a reader it seems a bit surprising to use a non-multithread-safe object on two threads, and is difficult/impossible to use without an Initialize method.

A couple of alternatives:
- Use rtc::Thread::Invoke to synchronously construct the object on a different thread (done here).
- Allocate memory for the object on the local thread then post an asynchronous task to construct it (using "placement new") on the target thread.

I think in WebRTC the first is most common. I haven't seen the second alternative used in our code but it seems valid.

Patch-set: 4
Reviewer: Gerrit User 5039 <5039@58829da1-049c-39fb-b951-ebdcd0984767>