Update patch set 35
Patch Set 35:

> Patch Set 35:
> 
> (1 comment)
> 
> > Patch Set 32:
> > 
> > (2 comments)
> > 
> > > Patch Set 32:
> > > 
> > > It looks like there might still be a launch review necessary in Chrome.
> > > Could you please add a build flag that needs to be enabled in order for h264 simulcast to be enabled? Then we can land this and only flip that switch when chrome is ready.
> > > Would rather not force you to keep rebasing every other day until then :(
> > 
> > What would I do then without my daily rebase?? ;)
> > 
> > The changes due to simulcast on h264 will still be there, as it replaces the original encode impl. So the only way of turning it down is by forcing the number of simulcast streams to 1 for h264.
> > 
> > Any bug present on the current patch may affect vp8 simulcast and h264 non-simulcast (even with the flag disabled), so not sure if it will protect much as currently there is no one enabling simulcast with h264.
> 
> I hear you, but I think this will still be needed.
> A flag to remove it from IsCodecBlacklistedForSimulcast() in webrtcvideoengine.cc should be fine.
> Then I'll stop nittin', promise!

No prob, just wanted to ensure we all understand the risks, I would not want to be blamed by hundreds of angry webrtc developers later if something breaks.. ;)

How about using a FieldTrial? Using a define is fine for me also, but I lack the knowledge to modify all the flags in the gn/ninja files to enable it on compile time. Also, that could allow us to use in chrome for testing.

Patch-set: 35
Reviewer: Gerrit User 5963 <5963@58829da1-049c-39fb-b951-ebdcd0984767>