Update patch set 37
Patch Set 37:

sing a FieldTrial? Using a define is fine for me also, but I lack the knowledge to modify all the flags in the gn/ninja files to enable it on compile time. Also, that could allow us to use in chrome for testing.
> 
> Sure, I'm fine with that.
Done!

Patch-set: 37
Reviewer: Gerrit User 5963 <5963@58829da1-049c-39fb-b951-ebdcd0984767>