Update patch set 13
Patch Set 13:

I just realized, I think we may have a problem... We store the layers in order of increasing quality, but the example in the spec (https://www.w3.org/TR/webrtc/#simulcast-example) does it in the reverse order. Same goes for the "a=simulcast" attribute we don't implement yet; the RIDs are listed in order of decreasing preference: https://tools.ietf.org/html/draft-ietf-mmusic-sdp-simulcast-12#section-5.2

This would also be relevant for the "active" flag, processed in UpdateSendState. If an application wants to use this to deactivate the highest quality encoding, it needs to know whether to turn off the first or last encoding.

This really is non-standard territory, though; the problem only exists due to our partial implementation of the API, where the only thing configurable (until now) is "number of layers" and we populate the default properties based on order. If we didn't have any built-in defaults, and required applications to explicitly set "scaleResolutionDownBy" in order to get different resolutions, then this wouldn't really be an issue. `sendEncodings: [{scaleResolutionDownBy: 2.0}, {}]` would be equivalent to `sendEncodings: [{}, {scaleResolutionDownBy: 2.0}]`; the application could pick whatever order it wants. (And `sendEncodings: [{}, {}]` would just pointlessly give you two identical encodings)

So, should we:
* Just skip straight to a full/standard implementation, avoiding this problem? I think this may be the least amount of effort in the end.
* Keep using the "increasing quality" order, and document it clearly in release notes/etc.
* Flip the order when translating between "encodings" and "simulcast_layers".

+Florent

Patch-set: 13
CC: Gerrit User 7036 <7036@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>