Delete FakeWebRtcVcmFactory::OnDestroyed method.
This was called by FakeWebRtcVideoCaptureModule's destructor.
However, since the factory keeps a reference counted pointer to
each FakeWebRtcVideoCaptureModule it has created, no module is
destroyed until the factory is destroyed. And at that point,
coordination is not needed and actually broken, since OnDestroyed
results in Release being called on an object halfway through the
destruction sequence.

Bug: webrtc:9405
Change-Id: I0cf3acb49b58e2b6b83344d840835f594418f7c7