Update patch set 10
Patch Set 10:

> It's hard to do it further down, as the decision to crop is made here in VideoStreamEncoder::ReconfigureEncoder().

I was thinking, if we delete the code here and pass on frames unchanged. Then if an encoder receives a frame of an odd size it doesn't like, that encoder can decide to ignore a few pixels at the edges, on a frame-by-frame basis.

> If we can do it cheaply, we should totally do that. I think, at the time this code was added WrapI420Buffer was not available. If it's possible to do it in this CL, let's do that. 

If we agree scaling isn't needed, a separate cl to replace scaling by cheap cropping makes sense to me. And we could try to move cropping closer to the encoder later on.

I also have a more general question on simulcast and sink wants. Say we have HD input and simulcast. Then available bandwidth drops, and it's decided (where, I don't know, but for simulcast I guess it's not the quality scaler?) to drop the highest resolution layer. Will we tell the videosource about that using sink wants, or will we keep passing high resolution frames to the encoder, and let the simulcast machinery scale down?

Patch-set: 10
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>