Update patch set 4
Patch Set 4:

It sounds good to me to use a struct rather than a lot of overloads.

A lot of the changes seems to be about changing so that the FEC Controller Factory is always injected, while the CL description does not mention this change directly. Maybe that could be a separate CL?

I'm not convinced that it's a good idea to make the construction of a call class more complicated since it's done so many time in many variations. The risk of creating a misconfigured end to end test will probably increase if we add more complex boilerplate into the construction of the tests. It might still be motivated, but I think it can be good to separate that topic from the struct vs overloads topic.

Patch-set: 4
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>