Update patch set 4
Patch Set 4:

I like this change! I'll hold off on the LGTM button until Magnus is happy with it, but we've needed to do something like this for a long time, so it's nice that someone finally stepped up to the plate so that I didn't have to. :-P

Patch-set: 4
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>