Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

The whole point of the caching is to avoid having to do a map lookup to find the correct histogram (except the first time). This CL makes it so that we have to do a map lookup every time anyway, because now we have one map per call site.

I think the histogram name argument is really intended to be a compile-time constant here. Is it possible to fix the problem at the call site instead? (What is the problematic call site, anyway?)

Patch-set: 3
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=-1