Update patch set 5
Patch Set 5: -Code-Review

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 2:
> > > 
> > > Hm, a +146 KB increase in libwebrtc.a looks pretty significant. Could you check with oprypin, kwiberg or nisse if this is a problem?
> > 
> > Yes, it looks pretty hefty: a 0.1% size increase for libjinge_peerconnection_so on arm32 and arm64.
> 
> I wonder if it's FieldTrialParameter that's costly?

It could be. It was introduced in an earlier CL without being used, but we said we won't start using it if it turns out to be too costly.

Sebastian, can you work with Karl to figure out if this makes sense?

Patch-set: 5
Reviewer: Gerrit User 5525 <5525@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=0