Update patch set 6
Patch Set 6:

I haven't found any way to reduce the size without also affecting the design in a significantly negative manner. I don't think that would be worth the maintainability tradeoff.

The option is to inject the BBR controller instead, that way the cost is only paid for those who are actually using it. This of course means a more complex design, especially if we want it to be used from the Java API.

WDYT?

Patch-set: 6
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>